Adds the new __main__.py of the tests "module".
This is basically just pretty glue logic - it pulls args and config as
defined and creates dictionaries from them for passage to the expected
runtime scripts.

For testing purposes I created dummy versions of the test scripts
themselves. These are non-functional and will need to be replaced.

Also committing in the config file.

A task should be added to the Testing Overhaul project to update
TESTDOCS if it does not already exist.

Signed-off-by: zadammac <zachary.j.adam@gmail.com>